src/test/javascript/spec/app/entities/address-manage/address-manage-update.component.spec.ts
/* tslint:disable max-line-length */
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { Observable, of } from 'rxjs';

import { ELearningTestModule } from '../../../test.module';
import { AddressManageUpdateComponent } from 'app/entities/address-manage/address-manage-update.component';
import { AddressManageService } from 'app/entities/address-manage/address-manage.service';
import { AddressManage } from 'app/shared/model/address-manage.model';

describe('Component Tests', () => {
    describe('AddressManage Management Update Component', () => {
        let comp: AddressManageUpdateComponent;
        let fixture: ComponentFixture<AddressManageUpdateComponent>;
        let service: AddressManageService;

        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [ELearningTestModule],
                declarations: [AddressManageUpdateComponent]
            })
                .overrideTemplate(AddressManageUpdateComponent, '')
                .compileComponents();

            fixture = TestBed.createComponent(AddressManageUpdateComponent);
            comp = fixture.componentInstance;
            service = fixture.debugElement.injector.get(AddressManageService);
        });

        describe('save', () => {
            it(
                'Should call update service on save for existing entity',
                fakeAsync(() => {
                    // GIVEN
                    const entity = new AddressManage(123);
                    spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
                    comp.address = entity;
                    // WHEN
                    comp.save();
                    tick(); // simulate async

                    // THEN
                    expect(service.update).toHaveBeenCalledWith(entity);
                    expect(comp.isSaving).toEqual(false);
                })
            );

            it(
                'Should call create service on save for new entity',
                fakeAsync(() => {
                    // GIVEN
                    const entity = new AddressManage();
                    spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
                    comp.address = entity;
                    // WHEN
                    comp.save();
                    tick(); // simulate async

                    // THEN
                    expect(service.create).toHaveBeenCalledWith(entity);
                    expect(comp.isSaving).toEqual(false);
                })
            );
        });
    });
});
